YUK-16005 Include multiple E2E requests/responses in a single packet Moved status from RfnDeviceResult to RfnCommandResult Fixed c1sx_display::cycle_delay warning

+++ yukon-server/RTDB/dev_rfn410centron.cpp
@@ -112,7 +112,7 @@ auto Rfn410CentronDevice::getDisplayConfigItems(const Config::DeviceConfigSPtr &
                 return getConfigDataEnum(deviceConfig, configKey, Cti::Config::displayItemMap);
             });
 
-    config.cycle_delay    = getConfigData<long, 0x00, 0x0f>(deviceConfig, Config::RfnStrings::LcdCycleTime);
+    config.cycle_delay    = getConfigData<uint8_t, 0x00, 0x0f>(deviceConfig, Config::RfnStrings::LcdCycleTime);
     config.display_digits = getConfigDataEnum(deviceConfig, Config::RfnStrings::DisplayDigits, displayDigitLookup);
 
     return config;
